+++ crates/test-programs/artifacts/Cargo.toml
@@ -16,7 +16,6 @@ wasmtime = { workspace = true, features = ['incremental-cache', 'cranelift', 'co
 [build-dependencies]
 heck = { workspace = true }
 wit-component = { workspace = true }
-wasmparser = { workspace = true, features = ['features', 'validate', 'component-model'] }
 wat = { workspace = true }
 cargo_metadata = "0.19.2"
 wasmtime-test-util = { workspace = true, features = ['wast'] }
